requirements.txt
akshare>=1.11.0
pandas>=2.0.0
numpy>=1.24.0
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
tqdm>=4.65.0
